update debug system and refactor
